move components into the corresponding folder
