+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css'
-import TodoList from './TodoList'
-import AddTodoForm from './AddTodoForm'
+import TodoList from './components/TodoList'
+import AddTodoForm from './components/AddTodoForm'
 
 function App() {
   return (
